refactor: move todo schemas from shared package to backend-local
Todo tools are backend-only (suppressToolEvents: true, no frontend UI).
Follow the dispatch_agent pattern: define schemas locally in the tool
directory instead of the shared @omnicraft/tool-schemas package.
