+++ packages/tool-schemas/src/index.ts
@@ -9,10 +9,6 @@ export {
   RUN_COMMAND_MAX_TIMEOUT_MS,
   runCommandParametersSchema,
   searchFilesParametersSchema,
-  todoAppendParametersSchema,
-  todoClearParametersSchema,
-  todoListParametersSchema,
-  todoUpdateParametersSchema,
   webFetchParametersSchema,
   webFetchRawParametersSchema,
   webSearchParametersSchema,
@@ -33,9 +29,6 @@ export {
   readFileResultSchema,
   runCommandResultSchema,
   searchFilesResultSchema,
-  todoItemSchema,
-  todoResultSchema,
-  todoStatusSchema,
   toolFailureDataSchema,
   webFetchRawResultSchema,
   webFetchResultSchema,
