+++ packages/tool-schemas/src/registry.ts
@@ -9,7 +9,6 @@ import {
   readFileResultSchema,
   runCommandResultSchema,
   searchFilesResultSchema,
-  todoResultSchema,
   toolFailureDataSchema,
   webFetchRawResultSchema,
   webFetchResultSchema,
@@ -32,10 +31,6 @@ export const toolResultSchemas = {
   [TOOL_NAME.WEB_SEARCH]: webSearchResultSchema,
   [TOOL_NAME.LOAD_SKILL]: loadSkillResultSchema,
   [TOOL_NAME.ASK_USER]: askUserResultSchema,
-  [TOOL_NAME.TODO_APPEND]: todoResultSchema,
-  [TOOL_NAME.TODO_UPDATE]: todoResultSchema,
-  [TOOL_NAME.TODO_CLEAR]: todoResultSchema,
-  [TOOL_NAME.TODO_LIST]: todoResultSchema,
 } as const;
 
 /** Infer the success result data type for a given tool name. */
@@ -59,7 +54,6 @@ export const toolResultDataSchema = z.union([
   webSearchResultSchema,
   loadSkillResultSchema,
   askUserResultSchema,
-  todoResultSchema,
   toolFailureDataSchema,
 ]);
